requirements.txt
streamlit==1.35.0
boto3==1.34.100
plotly==5.20.0
pandas==2.2.0
python-dotenv==1.0.1
anthropic==0.25.0
tenacity==8.2.3
pydantic==2.7.0
requests==2.31.0
protobuf==3.20.3
